third_party/blink/web_tests/svg/dynamic-updates/SVGFETurbulenceElement-dom-stitchTiles-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'stitchTiles' attribute of the SVGFETurbulenceElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS turbulence.getAttribute('stitchTiles') is "stitch"
PASS turbulence.getAttribute('stitchTiles') is "noStitch"
PASS successfullyParsed is true

TEST COMPLETE
